operational/src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
#
aiohttp==3.7.2            # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
anyconfig==0.9.11         # via kedro
appdirs==1.4.4            # via black
argon2-cffi==20.1.0       # via notebook
arrow==0.17.0             # via jinja2-time
async-generator==1.10     # via nbclient
async-timeout==3.0.1      # via aiohttp
asyncio==3.4.3            # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
attrs==20.3.0             # via aiohttp, black, jsonschema, pytest
backcall==0.2.0           # via ipython
binaryornot==0.4.4        # via cookiecutter
black==v19.10b0           # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
bleach==3.2.1             # via nbconvert
cachetools==4.1.1         # via kedro
certifi==2020.6.20        # via requests
cffi==1.14.3              # via argon2-cffi
chardet==3.0.4            # via aiohttp, binaryornot, requests
click==7.1.2              # via black, cookiecutter, flask, kedro, pip-tools
cookiecutter==1.6.0       # via kedro
coverage==5.3.1           # via pytest-cov
decorator==4.4.2          # via ipython
defusedxml==0.6.0         # via nbconvert
deprecated==1.2.11        # via pygithub
entrypoints==0.3          # via nbconvert
flake8==3.8.4             # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
flask==1.1.2              # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
fsspec==0.6.3             # via kedro
future==0.18.2            # via cookiecutter
idna==2.10                # via requests, yarl
ipykernel==5.3.4          # via ipywidgets, jupyter, jupyter-console, notebook, qtconsole
ipython-genutils==0.2.0   # via jupyterlab, nbformat, notebook, qtconsole, traitlets
ipython==7.19.0           # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in, ipykernel, ipywidgets, jupyter-console
ipywidgets==7.5.1         # via jupyter
isort==4.3.21             # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
itsdangerous==1.1.0       # via flask
jedi==0.17.2              # via ipython
jinja2-time==0.2.0        # via cookiecutter
jinja2==2.11.2            # via cookiecutter, flask, jinja2-time, nbconvert, notebook
jmespath==0.10.0          # via kedro
joblib==0.17.0            # via scikit-learn
jsonschema==3.2.0         # via jupyterlab-launcher, nbformat
jupyter-client==6.1.7     # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in, ipykernel, jupyter-console, kedro, nbclient, notebook, qtconsole
jupyter-console==6.2.0    # via jupyter
jupyter-core==4.6.3       # via jupyter-client, nbconvert, nbformat, notebook, qtconsole
jupyter==1.0.0            # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
jupyterlab-launcher==0.10.5  # via jupyterlab
jupyterlab-pygments==0.1.2  # via nbconvert
jupyterlab==0.31.1        # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
kedro-docker==0.2.1       # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
kedro==0.16.6             # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in, kedro-docker
llvmlite==0.34.0          # via numba
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
mistune==0.8.4            # via nbconvert
more-itertools==8.6.0     # via pytest
multidict==5.1.0          # via aiohttp, yarl
nbclient==0.5.1           # via nbconvert
nbconvert==6.0.7          # via jupyter, notebook
nbformat==5.0.8           # via ipywidgets, nbclient, nbconvert, nbstripout, notebook
nbstripout==0.3.3         # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
nest-asyncio==1.4.2       # via nbclient
notebook==6.1.4           # via jupyter, jupyterlab, jupyterlab-launcher, widgetsnbextension
numba==0.51.2             # via shap
numpy==1.19.3             # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in, numba, pandas, scikit-learn, scipy, shap, xgboost
packaging==20.4           # via bleach, pytest
pandas==1.1.4             # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in, shap
pandocfilters==1.4.3      # via nbconvert
parso==0.7.1              # via jedi
pathspec==0.8.1           # via black
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==5.3.1          # via kedro
pluggy==0.13.1            # via kedro, pytest
poyo==0.5.0               # via cookiecutter
prometheus-client==0.8.0  # via notebook
prompt-toolkit==3.0.8     # via ipython, jupyter-console
ptyprocess==0.6.0         # via pexpect, terminado
py==1.10.0                # via pytest
pycodestyle==2.6.0        # via flake8
pycparser==2.20           # via cffi
pyflakes==2.2.0           # via flake8
pygithub==1.54.1          # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
pygments==2.7.2           # via ipython, jupyter-console, jupyterlab-pygments, nbconvert, qtconsole
pyjwt==1.7.1              # via pygithub
pyparsing==2.4.7          # via packaging
pyrsistent==0.17.3        # via jsonschema
pytest-cov==2.10.1        # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
pytest-mock==1.13.0       # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
pytest==5.4.3             # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in, pytest-cov, pytest-mock
python-dateutil==2.8.1    # via arrow, jupyter-client, pandas
python-json-logger==0.1.11  # via kedro
pytz==2020.4              # via pandas
pyyaml==5.3.1             # via kedro
pyzmq==19.0.2             # via jupyter-client, notebook, qtconsole
qtconsole==4.7.7          # via jupyter
qtpy==1.9.0               # via qtconsole
regex==2020.11.13         # via black
requests==2.24.0          # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in, cookiecutter, pygithub
scikit-learn==0.23.2      # via shap
scipy==1.5.4              # via scikit-learn, shap, xgboost
send2trash==1.5.0         # via notebook
shap==0.37.0              # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
six==1.15.0               # via argon2-cffi, bleach, jsonschema, packaging, pip-tools, python-dateutil
slicer==0.0.3             # via shap
terminado==0.9.1          # via notebook
testpath==0.4.4           # via nbconvert
threadpoolctl==2.1.0      # via scikit-learn
toml==0.10.2              # via black
toposort==1.5             # via kedro
tornado==6.1              # via ipykernel, jupyter-client, notebook, terminado
tqdm==4.53.0              # via shap
traitlets==5.0.5          # via ipykernel, ipython, ipywidgets, jupyter-client, jupyter-core, nbclient, nbconvert, nbformat, notebook, qtconsole
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via aiohttp
urllib3==1.25.11          # via requests
wcwidth==0.2.5            # via prompt-toolkit, pytest
webencodings==0.5.1       # via bleach
werkzeug==1.0.1           # via flask
wheel==0.32.2             # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
whichcraft==0.6.1         # via cookiecutter
widgetsnbextension==3.5.1  # via ipywidgets
wrapt==1.12.1             # via deprecated
xgboost==1.2.1            # via -r /home/lukas/pjs/operativelogik/op-logik/src/requirements.in
yarl==1.6.3               # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
